for qualified test..

+++ analyzer/include/types.h
@@ -37,7 +37,7 @@ typedef struct {
 typedef struct {
 	float vx, vy, vz; // vx vy vz
 	float theta, phi, psi;
-	float altitude;
+	int altitude;
 	bool isflying;
 } NAVDATA;
